Trying to use only the tsx files instead of first compiling to js
