+++ index.html
@@ -18,6 +18,6 @@
     import this file directly. Adjust the path if your `main.tsx` is placed
     elsewhere (e.g. "./main.tsx" or "/main.tsx").
   -->
-  <script type="module" src="./src/main.js"></script>
+  <script type="module" src="./src/main.tsx"></script>
 </body>
 </html>
