+++ src/ErrorBoundary.tsx
@@ -1,24 +1,29 @@
-"use strict";
 // import React from 'react';
+
 // interface ErrorBoundaryState {
 //   hasError: boolean;
 //   error?: Error;
 // }
+
 // interface ErrorBoundaryProps {
 //   children: React.ReactNode;
 //   fallback?: React.ReactNode;
 // }
+
 // export class ErrorBoundary extends React.Component<ErrorBoundaryProps, ErrorBoundaryState> {
 //   constructor(props: ErrorBoundaryProps) {
 //     super(props);
 //     this.state = { hasError: false };
 //   }
+
 //   static getDerivedStateFromError(error: Error): ErrorBoundaryState {
 //     return { hasError: true, error };
 //   }
+
 //   componentDidCatch(error: Error, errorInfo: React.ErrorInfo) {
 //     console.error('ErrorBoundary caught an error:', error, errorInfo);
 //   }
+
 //   render() {
 //     if (this.state.hasError) {
 //       return this.props.fallback || (
@@ -34,6 +39,7 @@
 //         </div>
 //       );
 //     }
+
 //     return this.props.children;
 //   }
 // }
